move DEFAULT_COIN_BALANCE to `fuels-test-helpers`
